Use newlib arc4random_buf for random fs device
