+++ kernel/libc/newlib/Makefile
@@ -13,8 +13,8 @@
 #
 
 OBJS = lock_common.o newlib_close.o newlib_env_lock.o newlib_environ.o \
-	newlib_execve.o \
-	newlib_exit.o newlib_fork.o newlib_fstat.o newlib_getpid.o \
+	newlib_execve.o newlib_exit.o \
+	newlib_fork.o newlib_fstat.o newlib_getentropy.o newlib_getpid.o \
 	newlib_gettimeofday.o newlib_isatty.o newlib_kill.o newlib_link.o \
 	newlib_lseek.o newlib_malloc.o newlib_open.o \
 	newlib_read.o newlib_sbrk.o newlib_stat.o newlib_times.o \
